+++ example/api.cpp.
@@ -1,4 +1,4 @@
-#include "co_context.hpp"
+#include "co_context/io_context.hpp"
 #include "co_context/lazy_io.hpp"
 #include "co_context/co/mutex.hpp"
 #include "co_context/co/semaphore.hpp"
