+++ include/co_context/eager_io.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "co_context.hpp"
+#include "co_context/io_context.hpp"
 
 namespace co_context {
