+++ include/co_context/lazy_io.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "co_context.hpp"
+#include "co_context/io_context.hpp"
 #include <cassert>
 #include <span>
 #include <chrono>
