+++ lib/co_context/co/mutex.cpp
@@ -1,5 +1,5 @@
 #include "co_context/co/mutex.hpp"
-#include "co_context.hpp"
+#include "co_context/io_context.hpp"
 #include <cassert>
 
 namespace co_context {
