+++ lib/co_context/co/semaphore.cpp
@@ -1,5 +1,5 @@
 #include "co_context/co/semaphore.hpp"
-#include "co_context.hpp"
+#include "co_context/io_context.hpp"
 #include "co_context/log/log.hpp"
 #include <cassert>
