+++ test/sem.cpp
@@ -6,7 +6,7 @@
 #include "co_context/co/mutex.hpp"
 #include <random>
 #include "co_context/co/semaphore.hpp"
-#include "co_context.hpp"
+#include "co_context/io_context.hpp"
 #include "co_context/task.hpp"
 
 using namespace co_context;
